requirements.txt
apache-airflow
kafka-python==2.0.2